mcp_server/requirements.txt
fastmcp>=2.3.4
python-arango>=8.1.6
python-dotenv>=1.1.0